refactor(security-hub): split SecurityPanelView into view + hooks
Pull the panel apart along the view/logic seam:

Views (sibling files):
- SectionPanel.tsx — animated section wrapper.
- PanelHeader.tsx — score gauge + strength chip + action line.
- SignersSection.tsx — signers + multichain accordion.
- SecurityChecksSection.tsx — security-checks accordion.

Business logic (hooks/ subfolder):
- usePanelHeader.ts — derives score/level/color/actionLine from results.
- useSignerRows.tsx — signer + multichain row bucketing and sorting.
- useSecurityChecks.tsx — moved here; per-check row builder.

Plus a constants.ts holding STRENGTH_DESCRIPTIONS / GRADE_BG_BY_STRENGTH /
ROW_STAGGER. SecurityPanelView.tsx drops from 421 → 58 lines and now
just composes the three sub-sections + skeleton fallback. Storybook
snapshot tests pass unchanged.
