+++ apps/web/src/features/spaces/components/SecurityHub/components/SecurityPanelView/hooks/useSecurityChecks.tsx
@@ -13,7 +13,7 @@ import {
   makeBuildCta,
   sortBySeverity,
   type SectionRow,
-} from './primitives'
+} from '../primitives'
 
 export type UseSecurityChecksResult = {
   isReady: boolean
